# Improve navigation performance, reduce bundle sizes