--- a/packages/app/next.config.js
+++ b/packages/app/next.config.js
@@ -80,11 +80,6 @@ const nextConfig = {
   async redirects() {
     return [
       {
-        source: '/landelijk',
-        destination: '/landelijk/vaccinaties',
-        permanent: false,
-      },
-      {
         source: '/actueel',
         destination: '/',
         permanent: false,
@@ -148,9 +143,15 @@ const nextConfig = {
       ['d3-array', '../../node_modules/d3-array'],
       ['d3-color', '../../node_modules/d3-scale/node_modules/d3-color'],
       ['d3-geo', '../../node_modules/d3-geo'],
-      ['d3-interpolate', '../../node_modules/d3-scale/node_modules/d3-interpolate'],
+      [
+        'd3-interpolate',
+        '../../node_modules/d3-scale/node_modules/d3-interpolate',
+      ],
       ['react-is', '../../node_modules/react-is'],
-      ['unist-util-visit-parents', '../../node_modules/unist-util-visit-parents',],
+      [
+        'unist-util-visit-parents',
+        '../../node_modules/unist-util-visit-parents',
+      ],
     ];
 
     duplicatePackageResolves.forEach(([packageName, resolvedPath]) => {