--- a/packages/app/src/components/error-boundary/logic/use-component-props-report.ts
+++ b/packages/app/src/components/error-boundary/logic/use-component-props-report.ts
@@ -73,7 +73,6 @@ export function useComponentPropsReport(
     try {
       return extractPropsAndFillReport(children, propsReportRef);
     } catch (e) {
-      console.error(e);
       return children;
     }
   };