--- a/packages/app/src/pages/internationaal/index.tsx
+++ b/packages/app/src/pages/internationaal/index.tsx
@@ -1,3 +1,5 @@
+import { useRouter } from 'next/router';
+import { useEffect } from 'react';
 import { InLayout } from '~/domain/layout/in-layout';
 import { Layout } from '~/domain/layout/layout';
 import { useIntl } from '~/intl';
@@ -7,6 +9,8 @@ import {
   StaticProps,
 } from '~/static-props/create-get-static-props';
 import { getLastGeneratedDate } from '~/static-props/get-data';
+import { useBreakpoints } from '~/utils/use-breakpoints';
+import { useReverseRouter } from '~/utils/use-reverse-router';
 
 export const getStaticProps = withFeatureNotFoundPage(
   'inHomePage',
@@ -18,6 +22,16 @@ export default function InternationalPage(
 ) {
   const intl = useIntl();
   const { lastGenerated } = props;
+
+  const router = useRouter();
+  const reverseRouter = useReverseRouter();
+  const breakpoints = useBreakpoints();
+
+  useEffect(() => {
+    if (breakpoints.md) {
+      router.replace(reverseRouter.in.positiefGetesteMensen());
+    }
+  }, [breakpoints.md, reverseRouter.in, router]);
 
   return (
     <Layout