Currently, when going to /landelijk (or veiligheidsregio/[code] and gemeente/[code]), you get redirected to the vaccinations page. On mobile the `menu=1` query parameter is added so you get to see the metrics menu, which is otherwise hidden. This means that on mobile, the full vaccinations page needs to be loaded before the menu is visible, which can be really slow on some older devices. 

This PR changes this logic a little bit. It will now only redirect on desktop, and mobile gets to see the actual /landelijk page, which contains _only_ the metrics menu now. When testing this in Chrome with 'Slow 3G' and a 6x CPU throttle, it saves a full second of loading time going from the homepage to the NL metrics menu (from 3.25s to 2.25s). Lighthouse performance scores improve from ±26 to ±73. Data savings go towards 500kB for these metric menu's.

I left the `menu=1` logic in there since it is actually still a clean way of opening the menu on mobile once you are on one of the metric pages. To be exact, it offers a clean way of _closing_ the menu again since the menu just needs to link to the same page without the menu query parameter, while we otherwise would need to add a context & hook to offer a closing method to the menu link list, which would need to make a button of the link linking to the current page..., etc.

While testing this solution I discovered that a bundle related to the choropleth was still included in the page. After some investigation I found out our ErrorBoundary naively imports the Choropleth, TimeseriesChart and StackedChart, _just to be able to check if a given component is one of these types_. This meant that regardless of the content of a page, as long as there was an ErrorBoundary on there, all three of these components were imported. I rewrote that logic so it doesn't need to actually import those components anymore, which shaved off at least 3kB on most pages, but up to 180kB on pages which did not use any of those components.

**Before**
```
Page                                                                                       Size     First Load JS
┌ ● /                                                                                      11.9 kB         517 kB
├   /_app                                                                                  0 B             248 kB
├ ○ /404                                                                                   809 B           284 kB
├ ● /500                                                                                   808 B           284 kB
├ ● /actueel/gemeente                                                                      4.62 kB         503 kB
├ ● /actueel/gemeente/[code] (158143 ms)                                                   2.88 kB         508 kB
├   ├ /nl/actueel/gemeente/GM1659 (932 ms)
├   ├ /nl/actueel/gemeente/GM0809 (928 ms)
├   ├ /nl/actueel/gemeente/GM0995 (893 ms)
├   ├ /nl/actueel/gemeente/GM0263 (890 ms)
├   ├ /nl/actueel/gemeente/GM0420 (862 ms)
├   ├ /nl/actueel/gemeente/GM1948 (859 ms)
├   ├ /nl/actueel/gemeente/GM0168 (751 ms)
├   └ [+345 more paths] (avg 441 ms)
├ ● /actueel/veiligheidsregio                                                              5.28 kB         503 kB
├ ● /actueel/veiligheidsregio/[code] (15212 ms)                                            2.81 kB         508 kB
├   ├ /nl/actueel/veiligheidsregio/VR04 (840 ms)
├   ├ /nl/actueel/veiligheidsregio/VR07 (782 ms)
├   ├ /nl/actueel/veiligheidsregio/VR05 (779 ms)
├   ├ /nl/actueel/veiligheidsregio/VR09 (752 ms)
├   ├ /nl/actueel/veiligheidsregio/VR03 (725 ms)
├   ├ /nl/actueel/veiligheidsregio/VR10 (725 ms)
├   ├ /nl/actueel/veiligheidsregio/VR11 (713 ms)
├   └ [+18 more paths] (avg 550 ms)
├ λ /api/choropleth/[...param]                                                             0 B             248 kB
├ λ /api/choropleth/konva-node.d                                                           0 B             248 kB
├ λ /api/choropleth/topojson-client.d                                                      0 B             248 kB
├ λ /api/choropleth/topology                                                               0 B             248 kB
├ λ /api/data/timeseries/__tests__/[...param]                                              0 B             248 kB
├ λ /api/data/timeseries/[...param]                                                        0 B             248 kB
├ λ /api/topo-json/[type]                                                                  0 B             248 kB
├ ● /artikelen                                                                             4.56 kB         288 kB
├ ● /artikelen/[slug] (25937 ms)                                                           2.87 kB         483 kB
├   ├ /en/artikelen/hoeveel-covid-19-patienten-liggen-op-de-ic (1198 ms)
├   ├ /en/artikelen/wat-is-het-covid-19-sterftecijfer-op-basis-van-doodsoorzaak (1176 ms)
├   ├ /en/artikelen/klopt-het-risiconiveau-in-mijn-regio-wel (1098 ms)
├   ├ /nl/artikelen/wat-zeggen-de-uitslagen-van-de-ggd-teststraten (965 ms)
├   ├ /nl/artikelen/hoe-weten-we-hoeveel-besmettelijke-mensen-er-zijn (880 ms)
├   ├ /nl/artikelen/wat-is-het-covid-19-sterftecijfer-op-basis-van-doodsoorzaak (867 ms)
├   ├ /nl/artikelen/hoeveel-covid-19-patienten-liggen-op-de-ic (730 ms)
├   └ [+47 more paths] (avg 405 ms)
├ ● /contact                                                                               2.76 kB         481 kB
├ ● /gemeente                                                                              4.62 kB         503 kB
├ ● /gemeente/[code] (58836 ms)                                                            4.18 kB         502 kB
├   ├ /nl/gemeente/GM1723 (668 ms)
├   ├ /nl/gemeente/GM0059 (638 ms)
├   ├ /nl/gemeente/GM1680 (625 ms)
├   ├ /nl/gemeente/GM0484 (616 ms)
├   ├ /nl/gemeente/GM0358 (586 ms)
├   ├ /nl/gemeente/GM0613 (565 ms)
├   ├ /nl/gemeente/GM0361 (543 ms)
├   └ [+345 more paths]
├ ● /gemeente/[code]/positief-geteste-mensen (280792 ms)                                   5.62 kB         525 kB
├   ├ /nl/gemeente/GM0197/positief-geteste-mensen (1483 ms)
├   ├ /nl/gemeente/GM1961/positief-geteste-mensen (1387 ms)
├   ├ /nl/gemeente/GM0059/positief-geteste-mensen (1365 ms)
├   ├ /nl/gemeente/GM1680/positief-geteste-mensen (1302 ms)
├   ├ /nl/gemeente/GM0484/positief-geteste-mensen (1266 ms)
├   ├ /nl/gemeente/GM0358/positief-geteste-mensen (1235 ms)
├   ├ /nl/gemeente/GM0840/positief-geteste-mensen (1207 ms)
├   └ [+345 more paths] (avg 787 ms)
├ ● /gemeente/[code]/rioolwater (91299 ms)                                                 6.35 kB         527 kB
├   ├ /nl/gemeente/GM0879/rioolwater (438 ms)
├   ├ /nl/gemeente/GM0173/rioolwater (433 ms)
├   ├ /nl/gemeente/GM0269/rioolwater (427 ms)
├   ├ /nl/gemeente/GM0981/rioolwater (422 ms)
├   ├ /nl/gemeente/GM0498/rioolwater (419 ms)
├   ├ /nl/gemeente/GM0501/rioolwater (416 ms)
├   ├ /nl/gemeente/GM0293/rioolwater (410 ms)
├   └ [+345 more paths]
├ ● /gemeente/[code]/sterfte (171338 ms)                                                   6.87 kB         522 kB
├   ├ /nl/gemeente/GM1954/sterfte (839 ms)
├   ├ /nl/gemeente/GM0513/sterfte (832 ms)
├   ├ /nl/gemeente/GM0141/sterfte (737 ms)
├   ├ /nl/gemeente/GM0512/sterfte (676 ms)
├   ├ /nl/gemeente/GM0059/sterfte (675 ms)
├   ├ /nl/gemeente/GM0518/sterfte (671 ms)
├   ├ /nl/gemeente/GM0856/sterfte (668 ms)
├   └ [+345 more paths] (avg 482 ms)
├ ● /gemeente/[code]/vaccinaties (84756 ms)                                                4.42 kB         533 kB
├   ├ /nl/gemeente/GM0114/vaccinaties (424 ms)
├   ├ /nl/gemeente/GM0770/vaccinaties (418 ms)
├   ├ /nl/gemeente/GM0716/vaccinaties (404 ms)
├   ├ /nl/gemeente/GM0183/vaccinaties (383 ms)
├   ├ /nl/gemeente/GM1731/vaccinaties (382 ms)
├   ├ /nl/gemeente/GM1730/vaccinaties (376 ms)
├   ├ /nl/gemeente/GM0356/vaccinaties (364 ms)
├   └ [+345 more paths]
├ ● /gemeente/[code]/ziekenhuis-opnames (190446 ms)                                        5.66 kB         525 kB
├   ├ /nl/gemeente/GM0034/ziekenhuis-opnames (1126 ms)
├   ├ /nl/gemeente/GM0361/ziekenhuis-opnames (985 ms)
├   ├ /nl/gemeente/GM0762/ziekenhuis-opnames (947 ms)
├   ├ /nl/gemeente/GM0770/ziekenhuis-opnames (946 ms)
├   ├ /nl/gemeente/GM0228/ziekenhuis-opnames (931 ms)
├   ├ /nl/gemeente/GM0772/ziekenhuis-opnames (923 ms)
├   ├ /nl/gemeente/GM0141/ziekenhuis-opnames (894 ms)
├   └ [+345 more paths] (avg 532 ms)
├ ● /internationaal                                                                        1.11 kB         490 kB
├ ● /internationaal/positief-geteste-mensen                                                15.4 kB         525 kB
├ ● /internationaal/varianten                                                              5.79 kB         523 kB
├ ● /landelijk                                                                             1.48 kB         490 kB
├ ● /landelijk/besmettelijke-mensen                                                        5.2 kB          511 kB
├ ● /landelijk/brononderzoek                                                               4.4 kB          520 kB
├ ● /landelijk/coronamelder                                                                5.4 kB          511 kB
├ ● /landelijk/gedrag                                                                      5.16 kB         530 kB
├ ● /landelijk/gehandicaptenzorg                                                           3.9 kB          513 kB
├ ● /landelijk/intensive-care-opnames                                                      6.55 kB         516 kB
├ ● /landelijk/maatregelen                                                                 1.65 kB         508 kB
├ ● /landelijk/positief-geteste-mensen                                                     7.15 kB         516 kB
├ ● /landelijk/reproductiegetal                                                            5.49 kB         511 kB
├ ● /landelijk/rioolwater                                                                  2.39 kB         517 kB
├ ● /landelijk/sterfte                                                                     6.27 kB         516 kB
├ ● /landelijk/thuiswonende-ouderen                                                        3.73 kB         513 kB
├ ● /landelijk/vaccinaties                                                                 9.64 kB         528 kB
├ ● /landelijk/varianten                                                                   1.89 kB         515 kB
├ ● /landelijk/verdenkingen-huisartsen                                                     5.63 kB         511 kB
├ ● /landelijk/verpleeghuiszorg                                                            3.94 kB         513 kB
├ ● /landelijk/ziekenhuis-opnames                                                          2.78 kB         516 kB
├ ● /over                                                                                  2.78 kB         481 kB
├ ● /over-risiconiveaus                                                                    1.14 kB         481 kB
├ λ /sitemap.xml                                                                           305 B           248 kB
├ ● /toegankelijkheid                                                                      2.77 kB         481 kB
├ ● /veelgestelde-vragen                                                                   1.21 kB         482 kB
├ ● /veiligheidsregio                                                                      5.28 kB         503 kB
├ ● /veiligheidsregio/[code] (5331 ms)                                                     4.15 kB         502 kB
├   ├ /nl/veiligheidsregio/VR01
├   ├ /nl/veiligheidsregio/VR02
├   ├ /nl/veiligheidsregio/VR03
├   └ [+22 more paths]
├ ● /veiligheidsregio/[code]/brononderzoek (15467 ms)                                      7.67 kB         533 kB
├   ├ /nl/veiligheidsregio/VR07/brononderzoek (904 ms)
├   ├ /nl/veiligheidsregio/VR19/brononderzoek (812 ms)
├   ├ /nl/veiligheidsregio/VR20/brononderzoek (791 ms)
├   ├ /nl/veiligheidsregio/VR18/brononderzoek (762 ms)
├   ├ /nl/veiligheidsregio/VR21/brononderzoek (755 ms)
├   ├ /nl/veiligheidsregio/VR23/brononderzoek (722 ms)
├   ├ /nl/veiligheidsregio/VR16/brononderzoek (708 ms)
├   └ [+18 more paths] (avg 556 ms)
├ ● /veiligheidsregio/[code]/gedrag (8428 ms)                                              4.24 kB         539 kB
├   ├ /nl/veiligheidsregio/VR03/gedrag (504 ms)
├   ├ /nl/veiligheidsregio/VR02/gedrag (495 ms)
├   ├ /nl/veiligheidsregio/VR06/gedrag (422 ms)
├   ├ /nl/veiligheidsregio/VR05/gedrag (413 ms)
├   ├ /nl/veiligheidsregio/VR04/gedrag (388 ms)
├   ├ /nl/veiligheidsregio/VR01/gedrag (384 ms)
├   ├ /nl/veiligheidsregio/VR09/gedrag (373 ms)
├   └ [+18 more paths] (avg 303 ms)
├ ● /veiligheidsregio/[code]/gehandicaptenzorg (19377 ms)                                  7.37 kB         523 kB
├   ├ /nl/veiligheidsregio/VR09/gehandicaptenzorg (959 ms)
├   ├ /nl/veiligheidsregio/VR13/gehandicaptenzorg (829 ms)
├   ├ /nl/veiligheidsregio/VR05/gehandicaptenzorg (828 ms)
├   ├ /nl/veiligheidsregio/VR14/gehandicaptenzorg (823 ms)
├   ├ /nl/veiligheidsregio/VR21/gehandicaptenzorg (821 ms)
├   ├ /nl/veiligheidsregio/VR06/gehandicaptenzorg (817 ms)
├   ├ /nl/veiligheidsregio/VR01/gehandicaptenzorg (799 ms)
├   └ [+18 more paths] (avg 750 ms)
├ ● /veiligheidsregio/[code]/maatregelen (6921 ms)                                         4.19 kB         520 kB
├   ├ /nl/veiligheidsregio/VR01/maatregelen (406 ms)
├   ├ /nl/veiligheidsregio/VR06/maatregelen (359 ms)
├   ├ /nl/veiligheidsregio/VR12/maatregelen (329 ms)
├   ├ /nl/veiligheidsregio/VR09/maatregelen (325 ms)
├   ├ /nl/veiligheidsregio/VR19/maatregelen (316 ms)
├   ├ /nl/veiligheidsregio/VR05/maatregelen (304 ms)
├   ├ /nl/veiligheidsregio/VR02/maatregelen
├   └ [+18 more paths]
├ ● /veiligheidsregio/[code]/positief-geteste-mensen (32855 ms)                            7.09 kB         526 kB
├   ├ /nl/veiligheidsregio/VR07/positief-geteste-mensen (1587 ms)
├   ├ /nl/veiligheidsregio/VR05/positief-geteste-mensen (1563 ms)
├   ├ /nl/veiligheidsregio/VR03/positief-geteste-mensen (1541 ms)
├   ├ /nl/veiligheidsregio/VR01/positief-geteste-mensen (1481 ms)
├   ├ /nl/veiligheidsregio/VR06/positief-geteste-mensen (1472 ms)
├   ├ /nl/veiligheidsregio/VR08/positief-geteste-mensen (1432 ms)
├   ├ /nl/veiligheidsregio/VR04/positief-geteste-mensen (1409 ms)
├   └ [+18 more paths] (avg 1243 ms)
├ ● /veiligheidsregio/[code]/rioolwater (14253 ms)                                         5.91 kB         526 kB
├   ├ /nl/veiligheidsregio/VR06/rioolwater (797 ms)
├   ├ /nl/veiligheidsregio/VR09/rioolwater (737 ms)
├   ├ /nl/veiligheidsregio/VR20/rioolwater (673 ms)
├   ├ /nl/veiligheidsregio/VR04/rioolwater (669 ms)
├   ├ /nl/veiligheidsregio/VR11/rioolwater (667 ms)
├   ├ /nl/veiligheidsregio/VR02/rioolwater (646 ms)
├   ├ /nl/veiligheidsregio/VR05/rioolwater (645 ms)
├   └ [+18 more paths] (avg 523 ms)
├ ● /veiligheidsregio/[code]/sterfte (15406 ms)                                            4.14 kB         524 kB
├   ├ /nl/veiligheidsregio/VR09/sterfte (769 ms)
├   ├ /nl/veiligheidsregio/VR02/sterfte (754 ms)
├   ├ /nl/veiligheidsregio/VR01/sterfte (744 ms)
├   ├ /nl/veiligheidsregio/VR03/sterfte (739 ms)
├   ├ /nl/veiligheidsregio/VR04/sterfte (690 ms)
├   ├ /nl/veiligheidsregio/VR08/sterfte (686 ms)
├   ├ /nl/veiligheidsregio/VR06/sterfte (676 ms)
├   └ [+18 more paths] (avg 575 ms)
├ ● /veiligheidsregio/[code]/thuiswonende-ouderen (21365 ms)                               7.17 kB         522 kB
├   ├ /nl/veiligheidsregio/VR08/thuiswonende-ouderen (1107 ms)
├   ├ /nl/veiligheidsregio/VR02/thuiswonende-ouderen (1099 ms)
├   ├ /nl/veiligheidsregio/VR17/thuiswonende-ouderen (1042 ms)
├   ├ /nl/veiligheidsregio/VR07/thuiswonende-ouderen (1024 ms)
├   ├ /nl/veiligheidsregio/VR03/thuiswonende-ouderen (1012 ms)
├   ├ /nl/veiligheidsregio/VR04/thuiswonende-ouderen (931 ms)
├   ├ /nl/veiligheidsregio/VR01/thuiswonende-ouderen (924 ms)
├   └ [+18 more paths] (avg 790 ms)
├ ● /veiligheidsregio/[code]/vaccinaties (10671 ms)                                        4.45 kB         533 kB
├   ├ /nl/veiligheidsregio/VR03/vaccinaties (591 ms)
├   ├ /nl/veiligheidsregio/VR01/vaccinaties (498 ms)
├   ├ /nl/veiligheidsregio/VR04/vaccinaties (496 ms)
├   ├ /nl/veiligheidsregio/VR21/vaccinaties (496 ms)
├   ├ /nl/veiligheidsregio/VR05/vaccinaties (489 ms)
├   ├ /nl/veiligheidsregio/VR02/vaccinaties (478 ms)
├   ├ /nl/veiligheidsregio/VR06/vaccinaties (469 ms)
├   └ [+18 more paths] (avg 397 ms)
├ ● /veiligheidsregio/[code]/verpleeghuiszorg (24045 ms)                                   7.38 kB         523 kB
├   ├ /nl/veiligheidsregio/VR09/verpleeghuiszorg (1111 ms)
├   ├ /nl/veiligheidsregio/VR12/verpleeghuiszorg (1096 ms)
├   ├ /nl/veiligheidsregio/VR25/verpleeghuiszorg (1040 ms)
├   ├ /nl/veiligheidsregio/VR08/verpleeghuiszorg (1037 ms)
├   ├ /nl/veiligheidsregio/VR24/verpleeghuiszorg (1018 ms)
├   ├ /nl/veiligheidsregio/VR23/verpleeghuiszorg (1016 ms)
├   ├ /nl/veiligheidsregio/VR06/verpleeghuiszorg (1014 ms)
├   └ [+18 more paths] (avg 929 ms)
├ ● /veiligheidsregio/[code]/ziekenhuis-opnames (16065 ms)                                 5.71 kB         525 kB
├   ├ /nl/veiligheidsregio/VR02/ziekenhuis-opnames (775 ms)
├   ├ /nl/veiligheidsregio/VR03/ziekenhuis-opnames (768 ms)
├   ├ /nl/veiligheidsregio/VR09/ziekenhuis-opnames (738 ms)
├   ├ /nl/veiligheidsregio/VR01/ziekenhuis-opnames (722 ms)
├   ├ /nl/veiligheidsregio/VR04/ziekenhuis-opnames (710 ms)
├   ├ /nl/veiligheidsregio/VR11/ziekenhuis-opnames (705 ms)
├   ├ /nl/veiligheidsregio/VR07/ziekenhuis-opnames (704 ms)
├   └ [+18 more paths] (avg 608 ms)
├ ● /verantwoording                                                                        1.29 kB         482 kB
└ ● /weekberichten/[slug] (13472 ms)                                                       4.05 kB         482 kB
    ├ /en/weekberichten/wekelijkse-update-coronacijfers-14-september
    ├ /nl/weekberichten/wekelijkse-update-coronacijfers-14-september
    ├ /en/weekberichten/wekelijkse-update-coronacijfers-7-september
    └ [+77 more paths]
+ First Load JS shared by all                                                              248 kB
  ├ chunks/framework-bbe16bb72ca1ef54.js                                                   42.6 kB
  ├ chunks/main-5d31c7f85d93bd09.js                                                        31.5 kB
  ├ chunks/pages/_app-ac18e684796cefb0.js                                                  171 kB
  ├ chunks/webpack-35e10115854aae51.js                                                     2.17 kB
  └ css/2fa5746e266317e3.css                                                               325 B
```

**After**
```
Page                                                                                                  Size     First Load JS
┌ ● /                                                                                                 11.9 kB         513 kB
├   /_app                                                                                             0 B             248 kB
├ ○ /404                                                                                              809 B           284 kB
├ ● /500                                                                                              808 B           284 kB
├ ● /actueel/gemeente                                                                                 3.82 kB         436 kB
├ ● /actueel/gemeente/[code] (179731 ms)                                                              2.89 kB         504 kB
├   ├ /nl/actueel/gemeente/GM0074 (1077 ms)
├   ├ /nl/actueel/gemeente/GM0879 (1002 ms)
…